Wiring diagrams for react that use Alan's Eyeball data as super- and as subsets
createFilesForReact.rb has been changed to create 2 filemanager.txt files, one
using the super- and one using the subset wiring diagrams as biological
information.

"BIO = {"Bio.txt"};
In here it is input the Biological Matrix: The n x n matrix is the
probability- adjacency matrix where for entry aij represents the probability
between 0 and 1 that an edge from xj to xi exists.
This matrix is obtained from information from any previous biological
information on the structure of the static network
This type of file is OPTIONAL."